Move execution helper methods to ExecutionUtils class, re-use execution utils for one-short worker execution
