+++ detekt/wrapper/src/main/java/io/buildfoundation/bazel/detekt/Application.java
@@ -7,6 +7,9 @@ import io.buildfoundation.bazel.detekt.stream.Streams;
 import io.buildfoundation.bazel.detekt.stream.WorkerStreams;
 import io.reactivex.rxjava3.core.Scheduler;
 
+import java.util.Arrays;
+import java.util.List;
+
 public interface Application {
 
     void run(String[] args);
@@ -25,14 +28,30 @@ public interface Application {
 
         @Override
         public void run(String[] args) {
-            // TODO: handle custom argument parsing when running as one-shot
-            ExecutableResult result = executable.execute(args);
+            List<String> rawArgs = Arrays.asList(args);
+            boolean shouldRunAsTestTarget = ExecutionUtils.shouldRunAsTestTarget(rawArgs);
+
+            // Extract the output path of the execution result from the arguments
+            String executionResultOutputPath = ExecutionUtils.getExecutionResultOutputPath(rawArgs);
+
+            // Sanitize the Detekt arguments
+            List<String> detektArgs = ExecutionUtils.sanitizeDetektArguments(rawArgs);
+            ExecutableResult result = executable.execute(detektArgs.toArray(new String[0]));
+            int statusCode = result.statusCode();
 
             if (result instanceof ExecutableResult.Failure) {
                 streams.error().println(result.output());
             }
 
-            platform.exit(result.statusCode());
+            // Write the execution result to a file
+            ExecutionUtils.writeExecutionResultToFile(statusCode, executionResultOutputPath);
+
+            // Force the exit code to be 0 if detekt is run as a test target
+            if (shouldRunAsTestTarget) {
+                statusCode = 0;
+            }
+
+            platform.exit(statusCode);
         }
     }
